+++ src/components/CardCarousel/CardCarousel.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import useStore from "@/pages/globalstores";
+import useStore from "@/db/globalstore";
 import {
   CardWrapper,
   IMGwrapper,
