+++ src/components/Counter/Counter.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import useStore from "@/pages/globalstores";
+import useStore from "@/db/globalstore";
 import useSWR from "swr";
 import { useEffect, useState } from "react";
