+++ src/pages/index.js
@@ -1,6 +1,6 @@
 import { useSession } from "next-auth/react";
 import { useRouter } from "next/router";
-import useStore from "./globalstores";
+import useStore from "../db/globalstore";
 import { useEffect, useState } from "react";
 import useSWR from "swr";
 // Layout & Content
